pyproject.toml
[build-system]
requires = [
    "setuptools>=42",
    "wheel"
]
build-backend = "setuptools.build_meta"

[project]
name = "plotter"
version = "1.0.0"
description = "A CLI tool for managing SVG files and paper sizes."
requires-python = ">=3.6"
authors = [
    { name = "Emile Aubley" }
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]
dependencies = [
    "typer",
    "questionary",
    "rich",
    "pyyaml"
]

[project.scripts]
plotter = "cli.commands:app"
